FIX: i and limit not the same type error

+++ srcs/str/ft_atoi_base.c
@@ -6,7 +6,7 @@
 /*   By: lilefebv <lilefebv@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/06 17:57:46 by lilefebv          #+#    #+#             */
-/*   Updated: 2025/02/05 16:34:14 by lilefebv         ###   ########lyon.fr   */
+/*   Updated: 2025/02/05 16:40:07 by lilefebv         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -63,7 +63,7 @@ int	ft_atoi_base_limited(const char *str, char *base, int limit)
 	i = 0;
 	res = 0;
 	base_len = ft_strlen(base);
-	while (ft_strchr(base, str[i]) && str[i] && i < limit)
+	while (ft_strchr(base, str[i]) && str[i] && i < (size_t)limit)
 	{
 		res = base_len * res + char_pos_int_str(base, str[i]);
 		i++;
